+++ consensus/dummy/consensus.go
@@ -10,7 +10,6 @@ import (
 
 	"github.com/ava-labs/avalanchego/vms/components/gas"
 	"github.com/ava-labs/libevm/common"
-	"github.com/ava-labs/libevm/consensus/misc/eip4844"
 	"github.com/ava-labs/libevm/core/state"
 	"github.com/ava-labs/libevm/core/types"
 	"github.com/ava-labs/libevm/trie"
@@ -24,17 +23,11 @@ import (
 )
 
 var (
-	errUnclesUnsupported                   = errors.New("uncles unsupported")
-	errExtDataGasUsedNil                   = errors.New("extDataGasUsed is nil")
-	errExtDataGasUsedTooLarge              = errors.New("extDataGasUsed is not uint64")
-	ErrInvalidBlockGasCost                 = errors.New("invalid blockGasCost")
-	errInvalidExtDataGasUsed               = errors.New("invalid extDataGasUsed")
-	errInvalidExcessBlobGasBeforeCancun    = errors.New("invalid excessBlobGas before cancun")
-	errInvalidBlobGasUsedBeforeCancun      = errors.New("invalid blobGasUsed before cancun")
-	errInvalidParentBeaconRootBeforeCancun = errors.New("invalid parentBeaconRoot before cancun")
-	errMissingParentBeaconRoot             = errors.New("header is missing beaconRoot")
-	errNonEmptyParentBeaconRoot            = errors.New("invalid non-empty parentBeaconRoot")
-	errBlobsNotEnabled                     = errors.New("blobs not enabled on avalanche networks")
+	errUnclesUnsupported      = errors.New("uncles unsupported")
+	errExtDataGasUsedNil      = errors.New("extDataGasUsed is nil")
+	errExtDataGasUsedTooLarge = errors.New("extDataGasUsed is not uint64")
+	ErrInvalidBlockGasCost    = errors.New("invalid blockGasCost")
+	errInvalidExtDataGasUsed  = errors.New("invalid extDataGasUsed")
 )
 
 type Mode struct {
@@ -199,31 +192,7 @@ func verifyHeader(chain consensus.ChainHeaderReader, header *types.Header, paren
 	if diff := new(big.Int).Sub(header.Number, parent.Number); diff.Cmp(big.NewInt(1)) != 0 {
 		return consensus.ErrInvalidNumber
 	}
-	// Verify the existence / non-existence of excessBlobGas
-	cancun := config.IsCancun(header.Number, header.Time)
-	if !cancun {
-		switch {
-		case header.ExcessBlobGas != nil:
-			return fmt.Errorf("%w: have %d, expected nil", errInvalidExcessBlobGasBeforeCancun, *header.ExcessBlobGas)
-		case header.BlobGasUsed != nil:
-			return fmt.Errorf("%w: have %d, expected nil", errInvalidBlobGasUsedBeforeCancun, *header.BlobGasUsed)
-		case header.ParentBeaconRoot != nil:
-			return fmt.Errorf("%w: have %#x, expected nil", errInvalidParentBeaconRootBeforeCancun, *header.ParentBeaconRoot)
-		}
-	} else {
-		if header.ParentBeaconRoot == nil {
-			return errMissingParentBeaconRoot
-		}
-		if *header.ParentBeaconRoot != (common.Hash{}) {
-			return fmt.Errorf("%w: have %#x, expected empty", errNonEmptyParentBeaconRoot, *header.ParentBeaconRoot)
-		}
-		if err := eip4844.VerifyEIP4844Header(parent, header); err != nil {
-			return err
-		}
-		if *header.BlobGasUsed > 0 { // VerifyEIP4844Header ensures BlobGasUsed is non-nil
-			return fmt.Errorf("%w: used %d blob gas, expected 0", errBlobsNotEnabled, *header.BlobGasUsed)
-		}
-	}
+
 	return nil
 }